% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{clean_sales_ts}
\alias{clean_sales_ts}
\title{Czyszczenie i przygotowanie danych}
\usage{
clean_sales_ts(df, agg_level = "day")
}
\arguments{
\item{df}{Dataframe}

\item{agg_level}{Poziom agregacji: "day", "week", "month"}
}
\value{
Wyczyszczony dataframe
}
\description{
Czyszczenie danych (obsługa braków, duplikatów, sortowanie, opcja agregacji czasowej).
}
