% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compute_sales_metrics}
\alias{compute_sales_metrics}
\title{Obliczanie kluczowych metryk biznesowych}
\usage{
compute_sales_metrics(df)
}
\arguments{
\item{df}{Wyczyszczony Dataframe}
}
\value{
Dataframe z metrykami
}
\description{
Oblicza sprzedaż całkowitą, przeciętną, średnią kroczącą, zmienność i udział promocji
}
