% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_management_summary}
\alias{create_management_summary}
\title{Generowanie podsumowania}
\usage{
create_management_summary(df)
}
\arguments{
\item{df}{Dataframe}
}
\value{
Lista z informacjami dla menedżera
}
\description{
Tworzy podsumowanie biznesowe (najlepszy/najgorszy sklep, rosnąca kategoria, itp.)
}
