% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_prognosis}
\alias{create_prognosis}
\title{Tworzenie prognoz z wykorzystaniem ARIMA i Prophet}
\usage{
create_prognosis(df, periods = 12)
}
\arguments{
\item{df}{Dataframe}

\item{periods}{Liczba okresów do prognozy}
}
\value{
Lista z modelami
}
\description{
Generuje prognozy na podstawie dotychczasowych trendów, używając ARIMA oraz pakietu prophet.
}
