% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{load_sales_data}
\alias{load_sales_data}
\title{Wczytanie danych sprzedażowych}
\usage{
load_sales_data(train_path, stores_path, holidays_path)
}
\arguments{
\item{train_path}{Ścieżka do pliku train.csv}

\item{stores_path}{Ścieżka do pliku stores.csv}

\item{holidays_path}{Ścieżka do pliku holidays_events.csv}
}
\value{
Dataframe (tibble)
}
\description{
Wczytuje dane w odpowiednim formacie, korzystając z tidyverse.
Łączy zbiory train, stores i opcjonalnie holidays.
}
