% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_sales_trends}
\alias{plot_sales_trends}
\title{Wizualizacja trendów sprzedaży}
\usage{
plot_sales_trends(df_metrics)
}
\arguments{
\item{df_metrics}{Dataframe z metrykami}
}
\value{
Obiekt ggplot
}
\description{
Tworzy wizualizacje danych sprzedażowych.
}
