% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sales_ts_logic}
\alias{sales_ts_logic}
\title{Funkcja wyższego rzędu (Logika Szeregów Czasowych)}
\usage{
sales_ts_logic(
  df,
  p_city = NULL,
  p_state = NULL,
  start_date = NULL,
  end_date = NULL
)
}
\arguments{
\item{df}{Dataframe}

\item{p_city}{Miasto}

\item{p_state}{Stan}

\item{start_date}{Data początkowa}

\item{end_date}{Data końcowa}
}
\value{
Lista zawierająca metryki i wykres
}
\description{
Łączy obliczanie metryk i wizualizację na wybranych metadanych i w zadanym czasie
}
