% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{validate_sales_ts}
\alias{validate_sales_ts}
\title{Walidacja jakości danych sprzedażowych}
\usage{
validate_sales_ts(df)
}
\arguments{
\item{df}{Dataframe z historią sprzedaży}
}
\value{
Dataframe z listą ostrzeżeń
}
\description{
Sprawdzenie jakości danych (braki, duplikaty, poprawne daty, zakres wartości, częstotliwość)
}
